+++ CHANGELOG.md
@@ -43,6 +43,7 @@ Line wrap the file at 100 chars.                                              Th
 - Don't try to replace WireGuard key if account has too many keys already.
 - Fix bogus update notification caused by an outdated cache.
 - Fix layout issues when showing messages in WireGuard key view.
+- Disable WireGuard protocol option if there's no WireGuard key.
 
 #### Windows
 - Fix regression due to which a TAP adapter issue was not given as the specific block reason when